Vereinheitlichung einiger Tabellen wegen Ownership der Referenz. Nomen- klaturkorrektur userid -> user_id.
EntityWriteLocks.

+++ src/main/java/de/holarse/backend/db/repositories/ApiUserRepository.java
@@ -3,7 +3,7 @@ package de.holarse.backend.db.repositories;
 import de.holarse.backend.db.ApiUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApiUserRepository extends JpaRepository<ApiUser, Long> {
+public interface ApiUserRepository extends JpaRepository<ApiUser, Integer> {
     
     public ApiUser findByLogin(final String login);
